+++ packages/x-charts-pro/src/internals/plugins/zoomGestures/index.ts
@@ -1,4 +1,5 @@
-export { usePanGesture, type UsePanGestureOptions } from './usePanGesture';
+export { useDragGesture, type UseDragGestureOptions } from './useDragGesture';
+export { useDragOnPressGesture, type UseDragOnPressGestureOptions } from './useDragOnPressGesture';
 export { useWheelGesture, type UseWheelGestureOptions } from './useWheelGesture';
 export { usePinchGesture, type UsePinchGestureOptions } from './usePinchGesture';
 export type { GestureInstance, PanGestureConfig, ChartPoint } from './zoomGestures.types';
